somewhat refactor tests
