+++ shared-libs/rust/session-manager/Cargo.toml
@@ -20,3 +20,4 @@ openssl.workspace = true
 env_logger = "^0"
 rand = "^0.8"
 dotenvy = "^0"
+eyre = "^0.6"
